Implementation of 'Provider' package
