+++ pubspec.yaml
@@ -16,6 +16,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  provider: ^3.1.0
 
 dev_dependencies:
   flutter_test:
